Migrate message star to new structure
