--- app/message-star/server/index.js
@@ -1,3 +1,0 @@
-import './settings';
-import './starMessage';
-import './startup/indexes';
