+++ server/api/methods/messages/starMessage.js
@@ -1,0 +1,7 @@
+import { Meteor } from 'meteor/meteor';
+
+import { starMessage } from '../../../services/messages/functions/starMessage';
+
+Meteor.methods({
+	starMessage,
+});
