+++ server/settings/settingfiles/index.ts
@@ -2,3 +2,4 @@ import './2fa.settings';
 import './api.settings';
 import './pinmessage.settings';
 import './snippetmessage.settings';
+import './starmessage.settings';
