+++ server/startup/functions/starmessage.indexes.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 
-import { Messages } from '../../../models';
+import { Messages } from '../../models';
 
 Meteor.startup(function() {
 	return Meteor.defer(function() {
